Split PermissionsTest into multiple files
